==> 1_detection/3_bing-2.ipynb <==
#requests bing imagery for positively identified farm sites

import pandas as pd
import geopandas as gpd
from shapely.geometry import mapping
from aerialImageRetrieval_2 import AerialImageRetrieval

#select country
countries = ['ALB', 'EPSG:6962'],['CYP', 'EPSG:6312'],['ESP', 'EPSG:2062'],['FRA', 'EPSG:2154'],['HRV', 'EPSG:3765'],['ITA', 'EPSG:7794'],['MLT', 'EPSG:3034'],['TUR', 'EPSG:5637'],['GRC', 'EPSG:2100']
n=countries[8][0]

#load csv containing coordinates of polygon corners
f = '../targets_2_simrdwn/'
file = f + n + '/' + n + '.csv'
targets = pd.read_csv(file, delimiter=',', header=0)
print(len(targets))

tl = targets.drop(targets.index[0])
tl = tl.iloc[::5]
print(len(tl))

br = targets.drop(targets.index[0:3])
br = br.iloc[::5] 
print(len(br))

# create lists
lat1_list=tl['y'].tolist()
lon1_list=tl['x'].tolist()
lat2_list=br['y'].tolist()
lon2_list=br['x'].tolist()

file = f + n + '/' + n + '_index.csv'
index = pd.read_csv(file, delimiter=',', header=None)

for i in range(0, len(tl)): #do 1000 at a time
    lat1=round(lat1_list[i],5)
    lon1=round(lon1_list[i],5)
    lat2=round(lat2_list[i],5)
    lon2=round(lon2_list[i],5)
    ID=index[0][i]
    
    imgretrieval = AerialImageRetrieval(n, ID, lat1, lon1, lat2, lon2)
    if imgretrieval.max_resolution_imagery_retrieval():
        print("Retrieved image:", lat1, lon1, lat2, lon2)
    else:
        print("Cannot retrieve image:", lat1, lon1, lat2, lon2)
