==> 1_detection/4_add_data.ipynb <==
import os
import pandas as pd
import geopandas as gpd

#SET COUNTRY/CRS <-------------------------------------------------------------------------------CHANGE
countries = ['ALB', 'EPSG:6962'],['ESP', 'EPSG:2062'],['CYP', 'EPSG:6312'],['FRA', 'EPSG:2154'],['HRV', 'EPSG:3765'],['ITA', 'EPSG:7794'],['MLT', 'EPSG:3034'], ['TUR', 'EPSG:5637'],['GRC', 'EPSG:2100']
countries = ['ESP', 'EPSG:2062'],['ESP', 'EPSG:2062']

#load images to obtain filenames
folder = '../targets_2_simrdwn/1_SIMRDWN_final/annotated/'
files = os.listdir(folder)
files = pd.DataFrame(files)
print(len(files), 'annotated images')

def add(n):
    #load country geojson file
    f = '../targets_2_simrdwn/3_new/' + n[0] + '/'
    pts = gpd.read_file(f + n[0] + '_simrdwn_farm_pts.geojson')
    print(len(pts), n[0])

    #extract farm ID and country data from filenames
    farm_list, country_list = [], []
    for i in range(len(files)):
        farm_id = int(files[0][i].split(',')[2])
        farm_list.append(farm_id)
        #country = (files[0][i].split("'")[1])
        country = (files[0][i].split(',')[1]).split('_')[1]
        country_list.append(country)

    #create dataframe of filenames data
    data = pd.DataFrame(list(zip(farm_list, country_list, files[0])), columns =['farm', 'country', 'filename']) 
    data = data[data['country']==n[0]] #filter by country
    data = data.sort_values(by=['farm']) #sort by increasing
    print(len(data))

    #add filename to geojson file
    pts['filename']=data['filename'].to_list()
    pts.to_file(f + n[0] + '_simrdwn_farm_pts_v2.geojson', driver='GeoJSON')
    
for i in countries:
    add(i)
