==> 1_detection/6_missing_imagery.ipynb <==
import os
import pandas as pd
import geopandas as gpd
from shapely.geometry import Point
from geopandas.tools import sjoin
import matplotlib.pyplot as plt

#SET COUNTRY <----------------------------------------------------------------------------------CHANGE
#n = 'TUR' 
#n = 'GRC'
# n = 'ESP'
# nn = n + '_v2'
#n = 'HRV' 
n = 'ITA' 
#n = 'CYP' 
#n = 'MLT' 
#n = 'FRA' 
#n = 'ALB' 

#SET CRS #<-------------------------------------------------------------------------------------CHANGE
#crs = 'EPSG:5637' #turkey
#crs = 'EPSG:2100' #greece
crs = 'EPSG:2062' #spain
#crs = 'EPSG:3765' #croatia
#crs = 'EPSG:7794' #italy
#crs = 'EPSG:6312' #cyprus
#crs = 'EPSG:3034' #malta
#crs = 'EPSG:2154' #france
#crs = 'EPSG:6962' #albania
wgs84= {'init' :'EPSG:4326'}

#load SAR predictions
#pts = gpd.read_file('../targets_0_sar/' + n + '/sar_' + nn + '.geojson')

pts1 = gpd.read_file('../targets_0_sar/' + n + '/sar_' + n + '_e_v2.geojson')
pts2 = gpd.read_file('../targets_0_sar/' + n + '/sar_' + n + '_w_v2.geojson')
pts = pd.concat([pts1, pts2])

pts = pts.to_crs({'init': crs})

print(len(pts), 'total SAR predictions')

#load folder of missing images
#f = 'ESP_2020_05_31'
#f = 'ITA_2020_06_01'
folder = '/Users/Zack/0_thesis_bing/imagery_' + n + '_missing/'

files = os.listdir(folder)
print(len(files), 'acquisitions (', round(len(files)/len(pts),2), 'percent )')

#extract coordinates
x_list, y_list = [], []
for i in range(len(files)):
    img = files[i]
    y1 = float(img.split(',')[1])
    x1 = float(img.split(',')[2])
    y2 = float(img.split(',')[3])
    x2 = float((img.split(',')[4]).split(')')[0])
    #calculate centroid
    x = (x1+x2)/2
    y = (y1+y2)/2
    x_list.append(x)
    y_list.append(y)

#centroids of missing locations 
geometry = [Point(i) for i in zip(x_list, y_list)]
missing = gpd.GeoDataFrame(geometry=geometry, crs=wgs84)
missing = missing.to_crs({'init': crs})
print(missing.crs)
#missing.head()

mask = gpd.GeoDataFrame(geometry = missing.buffer(10))
print(mask.crs)
print(len(mask))
mask.head()

clip = sjoin(pts, mask, how='inner', op='within')
clip=gpd.GeoDataFrame(geometry=geometry, crs=wgs84)
clip=clip.to_crs({'init': crs})
print(len(clip))
clip.head()

buffer = gpd.GeoDataFrame(geometry = clip.buffer(500))
envelope = gpd.GeoDataFrame(geometry = buffer.envelope)
print(len(envelope))
envelope.head()

#missing imagery
envelope2 = envelope.copy()
envelope2['Dissolve'] = 0
envelope2=envelope2.dissolve('Dissolve')

# f = '../targets_1_bing/missing_ITA/missing_' + n
# envelope2.to_file(f + '_dis.shp')
# envelope2['geometry'] = envelope2['geometry'].to_crs(epsg=4326)
# envelope2.to_file(f + '_dis.geojson', driver='GeoJSON')

# envelope.to_file('../missing_' + n + '.shp')
# envelope['geometry'] = envelope['geometry'].to_crs(epsg=4326)
# envelope.to_file('../missing_' + n + '.geojson', driver='GeoJSON')

clip.to_file(f + '_pts.shp')
clip['geometry'] = clip['geometry'].to_crs(epsg=4326)
clip.to_file(f + '_pts.geojson', driver='GeoJSON')

#acquired imagery
# envelope2 = envelope.copy()
# envelope2['Dissolve'] = 0
# envelope2=envelope2.dissolve('Dissolve')

# envelope2.to_file('../acquired_' + n + '_dis.shp')
# envelope2['geometry'] = envelope2['geometry'].to_crs(epsg=4326)
# envelope2.to_file('../acquired_' + n + '_dis.geojson', driver='GeoJSON')

# # clip.to_file('../acquired_' + n + '_pts.shp')
# # clip['geometry'] = clip['geometry'].to_crs(epsg=4326)
# # clip.to_file('../acquired_' + n + '_pts.geojson', driver='GeoJSON')
